refactor: extract stream_converter from step_adaptor, revert step error to None
Move ChatResponseChunk-to-AG-UI conversion into a standalone
stream_converter.py module with no dependency on the step adaptor.
Revert step processing errors to return None (logged only) since
a single step failure should not signal RunErrorEvent to the client.
